reorganized adapters so that they are all in one folder
