+++ code/app/src/androidTest/java/ca/ualberta/ishelf/myBooksFragmentTest.java
@@ -1,7 +1,5 @@
 package ca.ualberta.ishelf;
 
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
@@ -9,6 +7,8 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import ca.ualberta.ishelf.RecyclerAdapters.myBooksFragment;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
