+++ code/app/src/ca/ualberta/ishelf/BorrowFragment.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v4.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.BorrowFragment</li>
+<li>ca.ualberta.ishelf.RecyclerAdapters.BorrowFragment</li>
 </ul>
 </li>
 </ul>
