+++ code/app/src/ca/ualberta/ishelf/LORRecyclerViewAdapter.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v7.widget.RecyclerView.Adapter&lt;<a href="../../../ca/ualberta/ishelf/LORRecyclerViewAdapter.ViewHolder.html" title="class in ca.ualberta.ishelf">LORRecyclerViewAdapter.ViewHolder</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.LORRecyclerViewAdapter</li>
+<li>ca.ualberta.ishelf.RecyclerAdapters.LORRecyclerViewAdapter</li>
 </ul>
 </li>
 </ul>
@@ -179,7 +179,7 @@ extends android.support.v7.widget.RecyclerView.Adapter&lt;<a href="../../../ca/u
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/LORRecyclerViewAdapter.html#onBindViewHolder-ca.ualberta.ishelf.LORRecyclerViewAdapter.ViewHolder-int-">onBindViewHolder</a></span>(<a href="../../../ca/ualberta/ishelf/LORRecyclerViewAdapter.ViewHolder.html" title="class in ca.ualberta.ishelf">LORRecyclerViewAdapter.ViewHolder</a>&nbsp;holder,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/LORRecyclerViewAdapter.html#onBindViewHolder-ca.ualberta.ishelf.RecyclerAdapters.LORRecyclerViewAdapter.ViewHolder-int-">onBindViewHolder</a></span>(<a href="../../../ca/ualberta/ishelf/LORRecyclerViewAdapter.ViewHolder.html" title="class in ca.ualberta.ishelf">LORRecyclerViewAdapter.ViewHolder</a>&nbsp;holder,
                 int&nbsp;position)</code>
 <div class="block">We define the functions that are called when the buttons are pressed in
  our recycler view.</div>
@@ -262,7 +262,7 @@ public&nbsp;<a href="../../../ca/ualberta/ishelf/LORRecyclerViewAdapter.ViewHold
 </dl>
 </li>
 </ul>
-<a name="onBindViewHolder-ca.ualberta.ishelf.LORRecyclerViewAdapter.ViewHolder-int-">
+<a name="onBindViewHolder-ca.ualberta.ishelf.RecyclerAdapters.LORRecyclerViewAdapter.ViewHolder-int-">
 <!--   -->
 </a>
 <ul class="blockList">
