+++ code/app/src/ca/ualberta/ishelf/MyAdapter.ViewHolder.html
@@ -95,7 +95,7 @@
 <li>android.support.v7.widget.RecyclerView.ViewHolder</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.MyAdapter.ViewHolder</li>
+<li>ca.ualberta.ishelf.RecyclerAdapters.MyAdapter.ViewHolder</li>
 </ul>
 </li>
 </ul>
