+++ code/app/src/ca/ualberta/ishelf/MyAdapter.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v7.widget.RecyclerView.Adapter&lt;<a href="../../../ca/ualberta/ishelf/MyAdapter.ViewHolder.html" title="class in ca.ualberta.ishelf">MyAdapter.ViewHolder</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.MyAdapter</li>
+<li>ca.ualberta.ishelf.RecyclerAdapters.MyAdapter</li>
 </ul>
 </li>
 </ul>
@@ -187,7 +187,7 @@ extends android.support.v7.widget.RecyclerView.Adapter&lt;<a href="../../../ca/u
 </tr>
 <tr id="i2" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/MyAdapter.html#onBindViewHolder-ca.ualberta.ishelf.MyAdapter.ViewHolder-int-">onBindViewHolder</a></span>(<a href="../../../ca/ualberta/ishelf/MyAdapter.ViewHolder.html" title="class in ca.ualberta.ishelf">MyAdapter.ViewHolder</a>&nbsp;holder,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/MyAdapter.html#onBindViewHolder-ca.ualberta.ishelf.RecyclerAdapters.MyAdapter.ViewHolder-int-">onBindViewHolder</a></span>(<a href="../../../ca/ualberta/ishelf/MyAdapter.ViewHolder.html" title="class in ca.ualberta.ishelf">MyAdapter.ViewHolder</a>&nbsp;holder,
                 int&nbsp;position)</code>&nbsp;</td>
 </tr>
 <tr id="i3" class="rowColor">
@@ -277,7 +277,7 @@ public&nbsp;<a href="../../../ca/ualberta/ishelf/MyAdapter.ViewHolder.html" titl
 </dl>
 </li>
 </ul>
-<a name="onBindViewHolder-ca.ualberta.ishelf.MyAdapter.ViewHolder-int-">
+<a name="onBindViewHolder-ca.ualberta.ishelf.RecyclerAdapters.MyAdapter.ViewHolder-int-">
 <!--   -->
 </a>
 <ul class="blockList">
