+++ code/app/src/ca/ualberta/ishelf/RequestFragment.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v4.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.RequestFragment</li>
+<li>ca.ualberta.ishelf.RecyclerAdapters.RequestFragment</li>
 </ul>
 </li>
 </ul>
