+++ code/app/src/ca/ualberta/ishelf/myBooksFragment.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v4.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.myBooksFragment</li>
+<li>ca.ualberta.ishelf.RecyclerAdapters.myBooksFragment</li>
 </ul>
 </li>
 </ul>
