+++ code/app/src/main/java/ca/ualberta/ishelf/ListOfRequestsActivity.java
@@ -8,7 +8,6 @@ package ca.ualberta.ishelf;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
@@ -27,6 +26,8 @@ import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.UUID;
+
+import ca.ualberta.ishelf.RecyclerAdapters.LORRecyclerViewAdapter;
 
 /**
  * This class deals with accepting requests that other users have made for one of our books
@@ -125,7 +126,7 @@ public class ListOfRequestsActivity extends AppCompatActivity {
      *
      * @param position the position in the recycler view of the clicked button
      */
-    void locationButton(int position) {
+    public void locationButton(int position) {
         Log.d(TAG+" locationButton", "Called with " + position);
         Request selectedRequest = requests.get(position);
         // send the "selectedRequest" Request to the MapsActivity passed as an extra called "Request"
@@ -143,7 +144,7 @@ public class ListOfRequestsActivity extends AppCompatActivity {
      * @param position the position in the recycler view of the clicked button
      * @author : Randal Kimpinski
      */
-    void acceptRequest(int position){
+    public void acceptRequest(int position){
         Log.d(TAG+" acceptRequest", "Called with " + position);
         // Create Database object that we will use
         Database db = new Database(this);
@@ -245,7 +246,7 @@ public class ListOfRequestsActivity extends AppCompatActivity {
      * @param position the position in the recycler view of the clicked button
      * @author : Randal Kimpinski
      */
-    void declineRequest(int position) {
+    public void declineRequest(int position) {
         Log.d(TAG+" declineRequest", "Called with " + position);
         // Create Database object that we will use
         Database db = new Database(this);
