+++ code/app/src/main/java/ca/ualberta/ishelf/RecyclerAdapters/BorrowAdapter.java
@@ -1,4 +1,4 @@
-package ca.ualberta.ishelf;
+package ca.ualberta.ishelf.RecyclerAdapters;
 
 import android.app.Activity;
 import android.content.Context;
@@ -43,6 +43,11 @@ import java.io.IOException;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
+
+import ca.ualberta.ishelf.Book;
+import ca.ualberta.ishelf.BookProfileActivity;
+import ca.ualberta.ishelf.Database;
+import ca.ualberta.ishelf.R;
 
 /**
  * BorrowAdapter
