+++ code/app/src/main/java/ca/ualberta/ishelf/RecyclerAdapters/BorrowFragment.java
@@ -1,27 +1,17 @@
-package ca.ualberta.ishelf;
+package ca.ualberta.ishelf.RecyclerAdapters;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.os.Bundle;
-import android.os.Handler;
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.app.SearchManager;
 import android.util.Log;
 import android.widget.AdapterView;
 import android.widget.SearchView;
-import android.widget.SearchView.OnQueryTextListener;
-import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
@@ -35,17 +25,13 @@ import com.google.gson.reflect.TypeToken;
 import java.lang.reflect.Type;
 
 import android.widget.Spinner;
-import android.widget.Toast;
 
-import com.firebase.client.DataSnapshot;
-import com.firebase.client.Firebase;
-import com.firebase.client.FirebaseError;
-import com.firebase.client.ValueEventListener;
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-
-import java.lang.reflect.Type;
 import java.util.ArrayList;
+
+import ca.ualberta.ishelf.Book;
+import ca.ualberta.ishelf.Database;
+import ca.ualberta.ishelf.R;
+import ca.ualberta.ishelf.User;
 
 import static android.support.constraint.Constraints.TAG;
