+++ code/app/src/main/java/ca/ualberta/ishelf/RecyclerAdapters/LORRecyclerViewAdapter.java
@@ -4,7 +4,7 @@
  * March 10, 2019
  * @author : Randal Kimpinski
  */
-package ca.ualberta.ishelf;
+package ca.ualberta.ishelf.RecyclerAdapters;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -19,15 +19,11 @@ import android.widget.RatingBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.firebase.client.DataSnapshot;
-import com.firebase.client.Firebase;
-import com.firebase.client.FirebaseError;
-import com.firebase.client.ValueEventListener;
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-
-import java.lang.reflect.Type;
 import java.util.ArrayList;
+
+import ca.ualberta.ishelf.ListOfRequestsActivity;
+import ca.ualberta.ishelf.R;
+
 /**
 * This is the adapter for the List of requests for the recycler view
  * @author : Randal Kimpinski
