+++ code/app/src/main/java/ca/ualberta/ishelf/RecyclerAdapters/RequestFragment.java
@@ -1,13 +1,9 @@
-package ca.ualberta.ishelf;
+package ca.ualberta.ishelf.RecyclerAdapters;
 
 import android.content.Context;
-import android.content.DialogInterface;
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.os.SharedMemory;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -28,10 +24,11 @@ import com.google.gson.reflect.TypeToken;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Date;
-import java.util.UUID;
 
-import static android.support.constraint.Constraints.TAG;
+import ca.ualberta.ishelf.Book;
+import ca.ualberta.ishelf.Database;
+import ca.ualberta.ishelf.R;
+import ca.ualberta.ishelf.Request;
 
 /**
  * Requests Activity
